requirements.txt
Flask-SocketIO==5.3.6
mysql-connector-python==8.3.0
flask==3.0.2
